Add Cookies class to manage request cookies
Cookie management differs from sessions as cookies are not only a
key-value store but also take expiry times, paths, domains and other
security-related options. While we can reuse the same API from Session
for fetching and settings, the new Cookies class takes these extra
options when setting (with secure defaults provided).

As the new class uses PHP's setcookie() under the hood, it's a little
tricky to test as it relies on sending HTTP headers so this is a
blindspot at the moment.

An effort is made to ensure that cookies are always up-to-date even
without making a new request: e.g. setting a cookie in a request makes
it immediately available when typically it would require a second
request.