Remove Application as it obscures the lifecycle
The Application object makes it seem like there is a long-lived
application running that is serving multiple requests when this is not
true. Instead, rely on explicitly creating a request and response for
every request to drive home the fact that the entire codebase is loaded
and executed on each request.